% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosieve.R
\docType{package}
\name{cytosieve-package}
\alias{cytosieve-package}
\alias{cytosieve}
\title{cytosieve}
\description{
cytosieve provides functionality for in silico cytometric filtering of bulk sequencing data with transcriptional profiles.
}
\details{
############################
}
\seealso{
\url{https://github.com/colbyford/cytosieve}
}
\author{
Colby T. Ford, \email{colby.ford@uncc.edu} \url{https://orcid.org/0000-0002-7859-3622}
}
\keyword{bioinformatics}
\keyword{biology}
\keyword{cytology}
\keyword{genomics}
\keyword{phylogenetics}
