% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_reads}
\alias{filter_reads}
\title{Filter Reads}
\usage{
filter_reads(input_path, output_path, genes_to_find = NULL,
  eliminate_matches = TRUE, pct_variability = 0.1, paired = FALSE,
  input_r2_path = NULL, output_r2_path = NULL, verbose = TRUE)
}
\arguments{
\item{input_path}{String. Path to the input .fastq file.}

\item{output_path}{String. Desired path to the output filtered .fastq file.}

\item{genes_to_find}{XString set of gene sequences for which to search. Usually longer sequences than the individual reads. Usually from a FASTA file.}

\item{eliminate_matches}{Logical. Should the matches that are found be filtered out (TRUE) or should non-matched be filtered out (FALSE)? (Default: TRUE)}

\item{pct_variability}{Numerical. The amount of variability that will be allowed in considering a match. (Default: 0.10)}

\item{paired}{Logical. Should the filtering occur on both a paired forward (R1) and reverse (R2) read?
If TRUE, then reads that are filtered from the R1 file will also be removed from the R2 file.  (Default: FALSE)}

\item{input_r2_path}{String. Path to the input (reverse) .fastq file. (Optional, only used when paired = TRUE)}

\item{output_r2_path}{String. Desired path to the output (reverse) filtered .fastq file. (Optional, only used when paired = TRUE)}

\item{verbose}{Logical. Should the process print out which read and gene it is currently searching? (Default: TRUE)}
}
\description{
Filter out reads that have genes that are not expressed in a particular stage/cell type of interest.
Generates a FASTQ file by filtering out reads based on a reference set of gene sequences.
}
