% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{read_gff}
\alias{read_gff}
\title{Reading and Writing GFF Files}
\usage{
read_gff(path, na_strings = c(".", "?"))
}
\arguments{
\item{path}{String. Path to the input/output .gff file.}

\item{na_strings}{Vector of strings.
Which strings in the GFF file are missing values and should be converted to NAs. (Default: `c(".", "?")`)}

\item{append}{Logical. Should the process overwrite the existing file? (Default: FALSE)}
}
\description{
Read and write General Feature Format (a.k. gene-finding format, generic feature format, GFF) is a file format for describing genes and other genetic features.
}
\examples{
gff <- read_gff("folder/myfile.gff")
write_gff(gff_filtered, path = "folder/myfile_filtered.gff")
}
